pi-config/requirements.txt
paho-mqtt
paramiko
